requirements.txt
pandas
numpy
scikit-learn
sentence-transformers
nltk
spacy
umap-learn
plotly
streamlit
matplotlib
joblib
tqdm
sqlalchemy
